Created StackArray
The stack array is an entirely macro based "smart" array. The idea is to
replicate the functionality of a dynamic array, such as  c++ vector, but
without (necessarily) allocating the data memory on the heap.

One benefit of this is that we can avoid a poiner indirection because
the memory is part of the size of the anonymous struct. A downside is
that is needs to be statically sized, but that's just the way it is.