Expose struct Cache
Need to expose struct Cache so that it's possible to allocate on the
stack and initialise in-place. This means that we are allowing people to
tamper with the internal state, so OOP enthusiasts are going to
be apoplectic, but really, just because something's there doesn't mean
that you should play with it.